Add more confirmation dialogs

+++ src/frontend/app/shared/components/list/list-types/cf-spaces/cf-space-card/cf-space-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Observable ,  Subscription } from 'rxjs';
+import { Observable, Subscription } from 'rxjs';
 import { map, switchMap, tap } from 'rxjs/operators';
 
 import { ISpace } from '../../../../../../core/cf-api.types';
@@ -21,6 +21,8 @@ import { CfUserService } from '../../../../../data-services/cf-user.service';
 import { MetaCardMenuItem } from '../../../list-cards/meta-card/meta-card-base/meta-card.component';
 import { CardCell } from '../../../list.types';
 import { CurrentUserPermissions } from '../../../../../../core/current-user-permissions.config';
+import { ConfirmationDialogService } from '../../../../confirmation-dialog.service';
+import { ConfirmationDialogConfig } from '../../../../confirmation-dialog.config';
 
 @Component({
   selector: 'app-cf-space-card',
@@ -51,7 +53,8 @@ export class CfSpaceCardComponent extends CardCell<APIResource<ISpace>> implemen
     private entityServiceFactory: EntityServiceFactory,
     private store: Store<AppState>,
     private cfOrgService: CloudFoundryOrganizationService,
-    private currentUserPermissionsService: CurrentUserPermissionsService
+    private currentUserPermissionsService: CurrentUserPermissionsService,
+    private confirmDialog: ConfirmationDialogService
   ) {
     super();
   }
@@ -154,11 +157,19 @@ export class CfSpaceCardComponent extends CardCell<APIResource<ISpace>> implemen
   }
 
   delete = () => {
-    this.cfOrgService.deleteSpace(
-      this.spaceGuid,
-      this.orgGuid,
-      this.cfEndpointService.cfGuid
+    const confirmation = new ConfirmationDialogConfig(
+      'Delete Space',
+      `Are you sure you want to delete space '${this.row.entity.name}'?`,
+      'Delete',
+      true
     );
+    this.confirmDialog.open(confirmation, () => {
+      this.cfOrgService.deleteSpace(
+        this.spaceGuid,
+        this.orgGuid,
+        this.cfEndpointService.cfGuid
+      );
+    });
   }
 
   goToSummary = () => this.store.dispatch(new RouterNav({
